fix: change FormatInfo.fps from int to float
yt-dlp can return fractional fps values (e.g. 0.469 for storyboards,
23.976 for NTSC video). Changed the type from int | None to float | None
to handle these values correctly. All 77 tests still pass.

+++ backend/src/models/video.py
@@ -5,7 +5,7 @@ class FormatInfo(BaseModel):
     format_id: str
     ext: str
     resolution: str
-    fps: int | None
+    fps: float | None
     vcodec: str
     acodec: str
     filesize: int | None
